[Core] Add uint24_t (poor mans solution)
